requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
djangorestframework==3.12.4
graphviz==0.17
idna==3.3
psycopg2-binary==2.9.1
python-dotenv==0.20.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
drf-yasg==1.20.0
coreapi==2.3.3
